+++ src/main/java/com/concert_reservation/api/application/facade/PointFacade.java
@@ -15,24 +15,6 @@ import lombok.RequiredArgsConstructor;
 @Component
 @RequiredArgsConstructor
 public class PointFacade {
-    private final PointService pointService;
-
-
-
-    public PointResponse chargePoint(PointRequest pointRequest) {
-        PointCommand pointCommand = new PointCommand(null, pointRequest.getUserId(), pointRequest.getBalance(), pointRequest.getAmount(), pointRequest.getPaymentTime(), pointRequest.getPaymentMethod());
-        PointInfo pointInfo = pointService.chargePoint(pointCommand);
-        return PointResponse.from(pointInfo);
-    }
-
-    public PointResponse getPoint(Long pointId) {
-        PointInfo pointInfo = pointService.getPoint(pointId);
-        return PointResponse.from(pointInfo);
-    }
-
-    public PointResponse deductPoint(PointRequest pointRequest) {
-        return PointResponse.from(pointService.deductPoint(pointRequest.toCommand()));
-    }
 
 
 }
