+++ src/main/java/com/concert_reservation/api/business/repo/TokenRepository.java
@@ -24,4 +24,6 @@ public interface TokenRepository {
 	List<Token> findAllByTokenStatusAndExpirationAtBefore(TokenStatus tokenStatus, LocalDateTime now);
 
 	List<Token> saveAll(List<Token> tokensToBeExpired);
+
+	int countByTokenStatusAndWaitingAtBefore(TokenStatus tokenStatus, LocalDateTime waitingAt);
 }
