+++ src/main/java/com/concert_reservation/api/business/service/TokenService.java
@@ -8,6 +8,7 @@ import com.concert_reservation.api.business.model.dto.info.TokenInfo;
 public interface TokenService {
 	TokenInfo createToken(TokenCommand tokenCommand);
 	TokenInfo getToken(String concertCode);
+	TokenInfo getUserTokenInfo(String userId);
 
 	void scheduledUpdateTokenStatusToProcessing();
