+++ src/test/java/com/concert_reservation/api/integration/WaitingQueueIntegrationTest.java
@@ -1,0 +1,4 @@
+package com.concert_reservation.api.integration;
+
+public class WaitingQueueIntegrationTest {
+}
